Start Logicals lib
