+++ tests/lib/test_binary.py
